Change /movie URI and handle MovieNotFoundException
The /movie endpoint now uses the syntax of /movie/title.
For example, /movie/titanic instead of previous /movie?title=titanic

When FetchFromCSV doesn't find a movie, a MovieNotFoundException is thrown.
This is handled by the Controller and an error message and 404 status code is returned.
The Error class is a simple object that holds the message and status code.

Closes #9