Refactor FetchFromCSV and Controller
Removed repeated code in FetchFromCSV, now it only has the ability to return all movies.
In Controller, FetchFromCSV is called once and stored as a final variable to be reused,
this way it doesn't need to read from the csv file every time an endpoint is called.

I've experimented with this using breakpoints, and it seems that the csv is only read once,
and that every subsequent call to an endpoint is using the array of movies already in memory.
This speeds up the calls to the endpoints because we don't need to read the csv file everytime.