@GetMapping(value = "/award.year/{year1}-{year2}")
If i search for movies in range of certain years with format "/award.year/1927-1928", then i will get a range of movie from 1927 to 1928.
Now if i just try to search for a certain year "/award.year/1927", then it wont work unless i add the dash at the end "/award.year/1927-"

@GetMapping(value = { "/award.year/{year1}", "/award.year/{year1}/{year2}" })

This will pretty much do the same thing except the dash "-" will be replaced with a slash "/". "/award.year/1927/1928" which will give me a range of movie from 1927 to 1928.
Now if i just try to search for a certain year "/award.year/1927", then it works and it gives me the movies only in 1927 with no slash "/" required at the end

Still trying to figure out how we can use the dash but just wanted to put this out there just in case i do not figure it out in time.
Hopefully that all made sense